// fe-challenge/src/types/news.ts
export interface NewsArticle {
  source: {
    id: string | null;
    name: string;
  };
  author: string | null;
  title: string;
  description: string | null;
  url: string;
  urlToImage: string | null;
  publishedAt: string;
  content: string | null;
}

export interface NewsResponse {
  status: string;
  totalResults: number;
  articles: NewsArticle[];
}

export interface NewsFilters {
  q?: string;
  category?: string;
  sources?: string;
  country?: string;
  pageSize?: number;
  page?: number;
  sortBy?: string;
}